avance

+++ src/main/java/com/cibertec/security/repository/EstadoRepository.java
@@ -3,5 +3,5 @@ package com.cibertec.security.repository;
 import com.cibertec.security.entity.Estado;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstadoRepository extends JpaRepository<Estado, Long> {
+public interface EstadoRepository extends JpaRepository<Estado, Integer> {
 }
