+++ src/main/java/com/cibertec/security/repository/SucursalRepository.java
@@ -3,5 +3,5 @@ package com.cibertec.security.repository;
 import com.cibertec.security.entity.Sucursal;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SucursalRepository extends JpaRepository<Sucursal, Long> {
+public interface SucursalRepository extends JpaRepository<Sucursal, Integer> {
 }
